Operator support: Add more operators and tests
Add more operators, such as LIKE, NOT LIKE, =~ and !~. The latter are
for regular expression matches.

Also add support for all the array comparison operators. The array
operators are a bit special, since the ANY, ALL and NONE requires an
additional operator, such as ==, !=, >, etc. to match against the
individual keys in the array.

Add test cases for all the operators from ArangoDB documentation, to
keep the tests up to par with the docs.